+++ explcheck/src/explcheck-preprocessing.lua
@@ -11,6 +11,7 @@ local range_flags = ranges.range_flags
 local EXCLUSIVE = range_flags.EXCLUSIVE
 local INCLUSIVE = range_flags.INCLUSIVE
 local MAYBE_EMPTY = range_flags.MAYBE_EMPTY
+local FIRST_MAP_THEN_SUBTRACT = range_flags.FIRST_MAP_THEN_SUBTRACT
 
 local lpeg = require("lpeg")
 local B, Cmt, Cp, Ct, Cc, P, V = lpeg.B, lpeg.Cmt, lpeg.Cp, lpeg.Ct, lpeg.Cc, lpeg.P, lpeg.V
@@ -55,10 +56,13 @@ local function analyze_and_report_issues(states, file_number, options)
       if #span_range > 0 then
         if index % 2 == 1 then  -- chunk of text
           local chunk_text = content:sub(span_range:start(), span_range:stop())
-          if content_started or chunk_text:find("%S") ~= nil then
+          if chunk_text:find("%S") ~= nil then
             content_started = true
+            table.insert(transformed_text_table, chunk_text)
+          else
+            -- Skip all empty lines that only contain comments and leading spaces.
+            table.insert(numbers_of_bytes_removed, {transformed_index, #span_range})
           end
-          table.insert(transformed_text_table, chunk_text)
         else  -- comment
           local comment_text = content:sub(span_range:start(), span_range:stop())
           local comment_start, ignored_issues = lpeg.match(parsers.ignored_issues, comment_text)
@@ -119,13 +123,19 @@ local function analyze_and_report_issues(states, file_number, options)
   local transformed_content, map_back = strip_comments()
 
   -- Determine which parts of the input files contain expl3 code.
-  local expl_ranges = {}
+  local expl_ranges, transformed_expl_ranges, outer_expl_ranges = {}, {}, {}
   local input_ended = false
 
-  local function capture_range(should_skip, range_start, range_end)
+  local function capture_range(outer_range_start, should_skip, range_start, range_end, outer_range_end)
     if not should_skip then
-      local range = new_range(range_start, range_end, EXCLUSIVE, #transformed_content, map_back, #content)
+      local flags = range_end == #transformed_content and INCLUSIVE or EXCLUSIVE
+      local outer_flags = outer_range_end == #transformed_content and INCLUSIVE or (EXCLUSIVE + FIRST_MAP_THEN_SUBTRACT)
+      local range = new_range(range_start, range_end, flags, #transformed_content, map_back, #content)
+      local transformed_range = new_range(range_start, range_end, flags, #transformed_content)
+      local outer_range = new_range(outer_range_start, outer_range_end, outer_flags, #transformed_content, map_back, #content)
       table.insert(expl_ranges, range)
+      table.insert(transformed_expl_ranges, transformed_range)
+      table.insert(outer_expl_ranges, outer_range)
     end
   end
 
@@ -204,10 +214,12 @@ local function analyze_and_report_issues(states, file_number, options)
     "Root";
     Root = (
       (
-        V"FirstLineExplPart" / capture_range
+        Cp()
+        * V"FirstLineExplPart" / capture_range
       )^-1
       * (
         V"NonExplPart"
+        * Cp()
         * V"ExplPart" / capture_range
       )^0
       * V"NonExplPart"
@@ -264,7 +276,9 @@ local function analyze_and_report_issues(states, file_number, options)
         V"Head"
         * Cp()
         * V"HeadlessCloser"
+        * Cp()
         + Cp()
+        * Cp()
         * parsers.eof
       )
     ),
@@ -306,6 +320,16 @@ local function analyze_and_report_issues(states, file_number, options)
 
   -- If no expl3 parts were detected, decide whether no part or the whole input file is in expl3.
   if(#expl_ranges == 0 and #content > 0) then
+
+    -- Record the while input file as a single contiguous expl3 part.
+    local function record_whole_file()
+      local range = new_range(1, #content, INCLUSIVE, #content)
+      local transformed_range = new_range(1, #transformed_content, INCLUSIVE, #transformed_content)
+      table.insert(expl_ranges, range)
+      table.insert(transformed_expl_ranges, transformed_range)
+      table.insert(outer_expl_ranges, range)
+    end
+
     issues:ignore({identifier_prefix = 'e102', seen = true})
     if expl3_detection_strategy == "precision" or expl3_detection_strategy == "never" then
       -- Assume that no part of the input file is in expl3.
@@ -314,11 +338,9 @@ local function analyze_and_report_issues(states, file_number, options)
       if expl3_detection_strategy == "recall" then
         issues:add('w100', 'no standard delimiters')
       end
-      local range = new_range(1, #content, INCLUSIVE, #content)
-      table.insert(expl_ranges, range)
+      record_whole_file()
     elseif expl3_detection_strategy == "auto" then
-      -- Use context clues to determine whether no part or the whole
-      -- input file is in expl3.
+      -- Use context clues to determine whether no part or the whole input file is in expl3.
       local expl3like_material_ratio = 0
       if #content > 0 then
         expl3like_material_ratio = expl3like_material_bytes / #content
@@ -326,21 +348,23 @@ local function analyze_and_report_issues(states, file_number, options)
       if expl3like_material_count >= get_option('min_expl3like_material_count', options, pathname)
           or expl3like_material_ratio >= get_option('min_expl3like_material_ratio', options, pathname) then
         issues:add('w100', 'no standard delimiters')
-        local range = new_range(1, #content, INCLUSIVE, #content)
-        table.insert(expl_ranges, range)
+        record_whole_file()
       end
     else
       assert(false, 'Unknown strategy "' .. expl3_detection_strategy .. '"')
     end
   end
 
+  assert(#expl_ranges == #transformed_expl_ranges)
+  assert(#expl_ranges == #outer_expl_ranges)
+
   -- Check for overlong lines within the expl3 parts.
-  for _, expl_range in ipairs(expl_ranges) do
+  for _, expl_range in ipairs(transformed_expl_ranges) do
     local offset = expl_range:start() - 1
 
     local function line_too_long(range_start, range_end)
       local range = new_range(offset + range_start, offset + range_end, EXCLUSIVE, #transformed_content, map_back, #content)
-      issues:add('s103', 'line too long', range)
+      issues:add('s103', 'line too long', range, content:sub(range:start(), range:stop()))
     end
 
     local overline_lines_grammar = (
@@ -357,6 +381,7 @@ local function analyze_and_report_issues(states, file_number, options)
   -- Store the intermediate results of the analysis.
   results.line_starting_byte_numbers = line_starting_byte_numbers
   results.expl_ranges = expl_ranges
+  results.outer_expl_ranges = outer_expl_ranges
   results.seems_like_latex_style_file = seems_like_latex_style_file
 end